refactor(hooks): address PR review on commit 54ed369
Six concrete changes from the review on PR #305:

- Drop the \`PostBuildHook\` newtype wrapper around \`FilePath\` (review
  comment "newtype useless"). The value is just a path threaded into
  \`System.Process.proc\` — wrapping it in a positional newtype added a
  destructor and an import burden for zero type-safety. \`Maybe FilePath\`
  everywhere now.
- Drop the \`OperatorHookSettings\` sub-record on \`WebSettings\` (review
  comment "should be part of CISettings"). \`postBuildHook\` now lives
  directly on \`CISettings\` next to \`maxConcurrentBuilds\`,
  \`autoBuildNewBranches\`, and \`jobRetentionDays\`. \`CICommand\` carries a
  plain \`Maybe FilePath\` and \`webSettings.ciSettings.postBuildHook\` is
  the single read site for web mode.
- Move all post-build hook code — \`hookEnvVars\`, \`runHook\`,
  \`postBuildImpl\`, \`HookError\`, \`displayHookError\`, \`hookTimeoutMicros\` —
  out of \`Vira.CI.Pipeline.Implementation\` and into a new
  \`Vira.CI.Pipeline.PostBuildHook\` module (review comment "isolate in a
  PostBuildHook.hs for legibility").
- Revert \`Vira.CI.Configuration\` to its \`main\` state (review comment
  "why is this file changing?"). The \`applyConfigWithRunner\` /
  \`configInterpreter\` split existed only for the hooks-example test
  that the police step had already deleted.
- Drop the stray blank line in \`package.yaml\` between \`record-hasfield\`
  and \`safecopy\` (review comment "why this diff?"). Regenerate
  \`vira.cabal\` via hpack.
- Drop the \`vira.hs\`-scoped \`NOTE\` block in \`doc/guide/config.md\`
  pointing at the operator hook (review comment "Remove this"). The
  post-build hook is not a \`vira.hs\` concept.
