+++ packages/vira/src/Vira/App/Run.hs
@@ -37,7 +37,6 @@ import Vira.CI.AutoBuild qualified as AutoBuild
 import Vira.CI.Cleanup.Daemon qualified as CleanupDaemon
 import Vira.CI.Context (CIMode (..), ViraContext (..))
 import Vira.CI.Pipeline qualified as Pipeline
-import Vira.CI.Pipeline.Effect (PostBuildHook)
 import Vira.CI.Pipeline.Program qualified as Program
 import Vira.CI.Worker qualified as Worker
 import Vira.CI.Worker.Type qualified as Worker
@@ -88,7 +87,7 @@ runVira = do
         whenJust (importFile webSettings) $ \filePath -> do
           importFromFileOrStdin acid (Just filePath)
 
-        let postBuildHook = webSettings.hookSettings.postBuildHook
+        let postBuildHook = webSettings.ciSettings.postBuildHook
 
         startTime <- getCurrentTime
         instanceInfo <- getInstanceInfo
@@ -140,7 +139,7 @@ runVira = do
       putTextLn $ "Platform: " <> platform instanceInfo
       putTextLn $ "Schema version: " <> show viraDbVersion
 
-    runCI :: GlobalSettings -> Maybe FilePath -> CIMode -> Maybe PostBuildHook -> IO ()
+    runCI :: GlobalSettings -> Maybe FilePath -> CIMode -> Maybe FilePath -> IO ()
     runCI gs mDir ciMode postBuildHook = do
       dir <- maybe getCurrentDirectory makeAbsolute mDir
       -- Throw Terminated on Ctrl+C so Process.hs cleanup logic can terminate nix processes
